requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.49
python-dotenv==1.0.0
Werkzeug==2.3.7
pdfplumber==0.10.2
openai==1.3.5
pandas==2.1.1
matplotlib==3.8.0
weasyprint==60.1
openpyxl==3.1.2
mysqlclient==2.2.7
Flask-Migrate==4.1.0
